lab2/pum2/requirements.txt
Pillow>=9.0.1
numpy~=1.21.4
scikit-learn~=1.0.2
matplotlib~=3.5.1
